Fix render state provider
